review :ft_atoi ft_bzero ft_calloc ft_isalnum ft_isalpha ft_isascii ft_isdigit ft_isprint ft_itoa ft_memmove ft_memset ft_strlen ft_strlcat ft_tolower ft_toupper soping at ft_strlcpy

+++ ft_strrchr.c
@@ -6,7 +6,7 @@
 /*   By: oamkhou <oamkhou@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/10/18 11:04:29 by oamkhou           #+#    #+#             */
-/*   Updated: 2025/10/19 00:00:27 by oamkhou          ###   ########.fr       */
+/*   Updated: 2025/10/24 03:08:02 by oamkhou          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,9 +14,9 @@
 
 char	*ft_strrchr(const char *s, int c)
 {
-	size_t	s_len;
+	int	s_len;
 
-	s_len = ft_strlen(s);
+	s_len = (int)ft_strlen(s);
 	while (s_len >= 0)
 	{
 		if (s[s_len] == (char)c)
